chore: Add node 13.x to the CI matrix
This required renaming .eslintrc.js to .eslintrc.cjs in order to make
eslint require() the file correctly.

I also required making requisite changes to .editorconfig,
.gitattributes, etc to properly handle the cjs file ext.

Additionally, the coverage/ dir was added to ignorePatterns. It
happened to pass the lint, but there's no reason to do it.

Also added endOfLine: "auto" to .eslintrc.cjs to fix failing linting
on windows.