fix: Bug when encoding more than one left/right paren in string literal
Per the docs, String.replace(target,dest) will only replace the first
occurrence when target is a string. A regex is needed to do a global
search and replace.

One odd thing I found when testing is that these worked, and replaced
all occurrences:
   ret.replace(",", "%x2C")
   ret.replace(":", "%x3A")

But these did not, and requred a regex to get all occurrences:
  ret.replace("(", "%x28")
  ret.replace(")", "%x28")

I used a regex for all regardless.