fix: Fix bug stringifying number-like keys and values (#645)
A string like "1e 6" will currently be stringified as 1e+6, which is
ambiguous because it looks like a number literal and it will be parsed
as one. This patch fixes the stringify code so that these strings are
quoted. AQF strings have a similar issue and this patch addresses them
as well.